+++ directord/drivers/grpcd.py
@@ -1089,16 +1089,6 @@ class Driver(drivers.BaseDriver):
             raise
         return True
 
-    def key_generate(self, keys_dir, key_type):
-        """Generate certificate.
-
-        :param keys_dir: Full Directory path where a given key will be stored.
-        :type keys_dir: String
-        :param key_type: Key type to be generated.
-        :type key_type: String
-        """
-        pass
-
     def heartbeat_send(
         self, host_uptime=None, agent_uptime=None, version=None, driver=None
     ):
