+++ directord/tests/test_drivers_zmq.py
@@ -15,6 +15,7 @@
 import unittest
 
 from unittest.mock import ANY
+from unittest.mock import call
 from unittest.mock import patch
 
 import zmq
@@ -365,3 +366,87 @@ class TestDriverZMQ(tests.TestBase):
             self.assertEqual(driver.bind_address, "10.1.10.1")
         finally:
             driver.shutdown()
+
+    def test_key_generate(self):
+        with patch("io.open", create=True):
+            self.driver._key_generate("foo", "bar")
+
+    @patch("os.rename", autospec=True)
+    @patch("os.listdir", autospec=True)
+    def test_move_cetrificates_null(self, mock_listdir, mock_rename):
+        mock_listdir.return_value = ["item-one", "item-two"]
+        self.driver._move_certificates(directory="/test/path")
+        mock_rename.assert_not_called()
+
+    @patch("os.rename", autospec=True)
+    @patch("os.listdir", autospec=True)
+    def test_move_cetrificates_normal(self, mock_listdir, mock_rename):
+        mock_listdir.return_value = ["item-one.key", "item-two.key"]
+        self.driver._move_certificates(directory="/test/path")
+        mock_rename.assert_called_with(
+            "/test/path/item-two.key", "/test/path/item-two.key"
+        )
+
+        mock_rename.assert_has_calls(
+            [
+                call("/test/path/item-one.key", "/test/path/item-one.key"),
+                call("/test/path/item-two.key", "/test/path/item-two.key"),
+            ]
+        )
+
+    @patch("os.rename", autospec=True)
+    @patch("os.listdir", autospec=True)
+    def test_move_cetrificates_backup(self, mock_listdir, mock_rename):
+        mock_listdir.return_value = ["item-one.key", "item-two.key"]
+        self.driver._move_certificates(directory="/test/path", backup=True)
+        mock_rename.assert_has_calls(
+            [
+                call("/test/path/item-one.key", "/test/path/item-one.key.bak"),
+                call("/test/path/item-two.key", "/test/path/item-two.key.bak"),
+            ]
+        )
+
+    @patch("os.rename", autospec=True)
+    @patch("os.listdir", autospec=True)
+    def test_move_cetrificates_target_directory(
+        self, mock_listdir, mock_rename
+    ):
+        mock_listdir.return_value = ["item-one.key", "item-two.key"]
+        self.driver._move_certificates(
+            directory="/test/path", target_directory="/new/test/path"
+        )
+        mock_rename.assert_has_calls(
+            [
+                call("/test/path/item-one.key", "/new/test/path/item-one.key"),
+                call("/test/path/item-two.key", "/new/test/path/item-two.key"),
+            ]
+        )
+
+    @patch("os.rename", autospec=True)
+    @patch("os.listdir", autospec=True)
+    def test_move_cetrificates_normal_selective(
+        self, mock_listdir, mock_rename
+    ):
+        mock_listdir.return_value = ["item-one.test", "item-two.key"]
+        self.driver._move_certificates(directory="/test/path", suffix=".test")
+        mock_rename.assert_called_once_with(
+            "/test/path/item-one.test", "/test/path/item-one.test"
+        )
+
+    @patch("os.makedirs", autospec=True)
+    @patch("os.rename", autospec=True)
+    @patch("os.listdir", autospec=True)
+    def test_generate_certificates(
+        self, mock_listdir, mock_rename, mock_makedirs
+    ):
+        mock_listdir.return_value = ["item-one.test", "item-two.key"]
+        with patch("io.open", create=True):
+            self.driver._generate_certificates()
+
+        mock_makedirs.assert_has_calls(
+            [
+                call("/etc/directord/certificates", exist_ok=True),
+                call("/etc/directord/public_keys", exist_ok=True),
+                call("/etc/directord/private_keys", exist_ok=True),
+            ]
+        )
