+++ directord/tests/test_main.py
@@ -379,7 +379,6 @@ class TestMain(tests.TestBase):
                 "export_jobs": None,
                 "export_nodes": None,
                 "filter": None,
-                "generate_keys": False,
                 "job_info": None,
                 "job_port": 5555,
                 "analyze_job": None,
@@ -412,7 +411,6 @@ class TestMain(tests.TestBase):
                 "export_jobs": None,
                 "export_nodes": None,
                 "filter": None,
-                "generate_keys": False,
                 "job_info": None,
                 "job_port": 5555,
                 "analyze_job": None,
@@ -445,7 +443,6 @@ class TestMain(tests.TestBase):
                 "export_jobs": None,
                 "export_nodes": None,
                 "filter": None,
-                "generate_keys": False,
                 "job_info": None,
                 "job_port": 5555,
                 "analyze_job": None,
@@ -478,7 +475,6 @@ class TestMain(tests.TestBase):
                 "export_jobs": None,
                 "export_nodes": None,
                 "filter": None,
-                "generate_keys": False,
                 "job_info": None,
                 "job_port": 5555,
                 "analyze_job": None,
@@ -511,7 +507,6 @@ class TestMain(tests.TestBase):
                 "export_jobs": None,
                 "export_nodes": None,
                 "filter": None,
-                "generate_keys": False,
                 "job_info": "xxxx",
                 "job_port": 5555,
                 "analyze_job": None,
@@ -544,7 +539,6 @@ class TestMain(tests.TestBase):
                 "export_jobs": "xxxx",
                 "export_nodes": None,
                 "filter": None,
-                "generate_keys": False,
                 "job_info": None,
                 "job_port": 5555,
                 "analyze_job": None,
@@ -577,40 +571,6 @@ class TestMain(tests.TestBase):
                 "export_jobs": None,
                 "export_nodes": "xxxx",
                 "filter": None,
-                "generate_keys": False,
-                "job_info": None,
-                "job_port": 5555,
-                "analyze_job": None,
-                "analyze_parent": None,
-                "analyze_all": False,
-                "backend_port": 5556,
-                "heartbeat_interval": 60,
-                "identity": None,
-                "socket_group": "0",
-                "socket_path": "/var/run/directord.sock",
-                "cache_path": "/var/cache/directord",
-                "list_jobs": False,
-                "list_nodes": False,
-                "mode": "manage",
-                "purge_jobs": False,
-                "purge_nodes": False,
-            },
-        )
-
-    def test__args_manage_generate_keys(self):
-        args, _ = main._args(["manage", "--generate-keys"])
-        self.assertDictEqual(
-            vars(args),
-            {
-                "config_file": None,
-                "datastore": "memory",
-                "debug": False,
-                "driver": "zmq",
-                "dump_cache": False,
-                "export_jobs": None,
-                "export_nodes": None,
-                "filter": None,
-                "generate_keys": True,
                 "job_info": None,
                 "job_port": 5555,
                 "analyze_job": None,
@@ -852,7 +812,6 @@ class TestMain(tests.TestBase):
             "driver": "zmq",
             "export_jobs": None,
             "export_nodes": None,
-            "generate_keys": False,
             "job_info": None,
             "job_port": 5555,
             "backend_port": 5556,
